Add metadata validation level; restructure into three levels (#82, #80)
Splits the comparison library into three level-aligned modules:
- schema.rs: the dictionary itself (S## checks), moved out of lib.rs
- meta.rs:   column names + types (M01-M03), reads only the data schema
- data.rs:   values (D01 nulls-in-required), the only level that scans

lib.rs now holds the shared comparison vocabulary (Level, ColumnIssue,
IssueKind with code()/level(), CompareReport, CompareError) and the
compare() orchestrator both data levels delegate to. Issues now carry their
M/D code in text and JSON output.

CLI gains validate-meta and validate-data (replacing parquet validate);
parquet types becomes types parquet. validate-schema keeps its name.
